feat(constants): move to fed regtest constants unit tests class
